+++ src/routes/post.js
@@ -37,6 +37,9 @@ router.get("/all", async (req, res) => {
     }catch(e){
         result.message = e.message
     }finally{
+        if(client){
+            client.end() 
+        }
         res.send(result)
     }
 })
@@ -61,6 +64,9 @@ router.post("/", async (req, res) => {
     }catch(e){
         result.message = e.message
     }finally{
+        if(client){
+            client.end()
+        }
         res.send(result)
     }
 })
@@ -69,7 +75,7 @@ router.post("/", async (req, res) => {
 router.get('/:postIdx', async (req, res) => {
     const { postIdx } = req.params
     const sql = `
-        SELECT P.title, P.crerated_at,P.content, A.idx, A.nickname, PC.name
+        SELECT P.title, P.crerated_at,P.content,P.post_likes A.idx, A.nickname, PC.name
         FROM backend.post P 
         LEFT OUTER JOIN backend.account A ON P.account_idx = A.idx
         LEFT OUTER JOIN backend.post_category PC ON P.post_category_idx = PC.idx
@@ -91,6 +97,9 @@ router.get('/:postIdx', async (req, res) => {
     }catch(e){
         result.message = e.message
     }finally{
+        if(client){
+            client.end() 
+        }
         res.send(result)
     }
 })
@@ -98,7 +107,7 @@ router.get('/:postIdx', async (req, res) => {
 //게시글 수정
 router.put("/:postIdx", async (req, res) => {
     const { postIdx } = req.params
-    const sql = "UPDATE backend.post SET title=$1, content=$2,post_category_idx=$3 WHERE idx = $4"
+    const sql = "UPDATE backend.post SET title=$1, content=$2,post_category_idx=$3 WHERE idx = $4 AND account_idx = $5"
     const { title, content, category } = req.body
     const result = {
         "success" : false,
@@ -108,13 +117,16 @@ router.put("/:postIdx", async (req, res) => {
         validator.session(req.session.idx)
         validator.post({title:title, content:content})
 
-        await client.query(sql, [title, content, category, postIdx])
+        await client.query(sql, [title, content, category, postIdx, req.session.idx])
 
         result.success = true
 
     }catch(e){
         result.message = e.message
     }finally{
+        if(client){
+            client.end() 
+        }
         res.send(result)
     }
 })
@@ -122,7 +134,7 @@ router.put("/:postIdx", async (req, res) => {
 //게시글 삭제
 router.delete("/:postIdx", async (req, res) => {
     const { postIdx } = req.params
-    const sql = "DELETE FROM backend.post WHERE idx = $1"
+    const sql = "DELETE FROM backend.post WHERE idx = $1 AND account_idx = $2"
     const result = {
         "success" : false,
         "message": ""
@@ -130,15 +142,55 @@ router.delete("/:postIdx", async (req, res) => {
     try{
         validator.session(req.session.idx)
 
-        await client.query(sql, [postIdx])
+        await client.query(sql, [postIdx, req.session.idx])
         result.success = true 
 
     }catch(e){
         result.message = e.message
     }finally{
+        if(client){
+            client.end() 
+        }
         res.send(result)
     }
 })
+
+//게시물 좋아요
+router.put("/postIdx/likes", async (req, res) => {
+    const { postIdx } = req.params
+    const result = {
+        "success" : false,
+        "message": ""
+    }
+
+    try{
+        validator.session(req.session.idx)
+
+        await client.query('BEGIN')
+        const isLikers = await client.query('SELECT idx FROM backend.post_likes WHERE post_idx = $1 AND account_idx = $2', [postIdx,req.session.idx])
+        if(isLikers.rows){
+            await client.query('DELETE FROM backend.post_likes WHERE post_idx = $1 AND account_idx = $2', [postIdx, req.session.idx])
+            await client.query('UPDATE backend.post SET post_likes = post_likes - 1 WHERE idx = $1 AND account_idx = $2')
+
+            result.message = "좋아요 취소"
+        }else{
+            await client.query('INSERT INTO backend.post_likes(post_idx, account_idx) VALUES($1, $2)', [postIdx, req.session.idx])
+            await client.query('UPDATE backend.post SET post_likes = post_likes + 1 WHERE idx = $1 AND account_idx = $2')
+            
+            result.message = "좋아요 추가"
+        }
+        await client.query('COMMIT')
+        result.success = true
+    }catch(e){
+        await client.query('ROLLBACK')
+        result.message = e.message
+    }finally{
+        if(client){
+            client.end() 
+        }
+        res.send(result)
+    }
+} )
